+++ perun/logic/index.py
@@ -555,7 +555,7 @@ def register_in_index(index_filename, registered_file, registered_file_checksum,
     perun_log.info("'{}' successfully registered in minor version index".format(reg_rel_path))
 
 
-def remove_from_index(base_dir, minor_version, removed_file_generator, remove_all=False):
+def remove_from_index(base_dir, minor_version, removed_file_generator):
     """Removes stream of removed files from the index.
 
     Iterates through all of the removed files, and removes their partial/full occurence from the
@@ -565,10 +565,10 @@ def remove_from_index(base_dir, minor_version, removed_file_generator, remove_al
     :param str minor_version: sha-1 representation of the minor version of vcs (like e..g commit)
     :param generator removed_file_generator: generator of filenames, that will be removed from the
         tracking
-    :param bool remove_all: true if all of the entries should be removed
     """
     # Get directory and index
     _, minor_version_index = store.split_object_name(base_dir, minor_version)
+    removed_profile_number = len(removed_file_generator)
 
     if not os.path.exists(minor_version_index):
         raise EntryNotFoundException("", "empty index")
@@ -580,7 +580,7 @@ def remove_from_index(base_dir, minor_version, removed_file_generator, remove_al
         all_entries.sort(key=lambda unsorted_entry: unsorted_entry.offset)
         removed_entries = []
 
-        for removed_file in removed_file_generator:
+        for i, removed_file in enumerate(removed_file_generator):
             def lookup_function(entry):
                 """Helper lookup function according to the type of the removed file"""
                 if store.is_sha1(removed_file):
@@ -588,15 +588,14 @@ def remove_from_index(base_dir, minor_version, removed_file_generator, remove_al
                 else:
                     return entry.path == removed_file
 
-            if remove_all:
-                removed_entries.append(
-                    lookup_all_entries_within_index(index_handle, lookup_function)
-                )
-            else:
-                removed_entries.extend([
-                    lookup_entry_within_index(index_handle, lookup_function, removed_file)
-                ])
-            perun_log.info("deregistered: {}".format(removed_file))
+            found_entry = lookup_entry_within_index(index_handle, lookup_function, removed_file)
+            removed_entries.append(found_entry)
+
+            perun_log.info("{}/{} deregistered {} from index".format(
+                helpers.format_counter_number(i+1, removed_profile_number),
+                removed_profile_number,
+                perun_log.in_color(found_entry.path, 'grey')
+            ))
 
         # Update number of entries
         index_handle.seek(INDEX_NUMBER_OF_ENTRIES_OFFSET)
@@ -609,6 +608,14 @@ def remove_from_index(base_dir, minor_version, removed_file_generator, remove_al
             entry.write_to(index_handle)
 
         index_handle.truncate()
+    if removed_profile_number:
+        result_string = perun_log.in_color("{}".format(
+            helpers.str_to_plural(removed_profile_number, "profile")
+        ), 'white', 'bold')
+        index_sha = perun_log.in_color(minor_version, 'green')
+        perun_log.info("successfully deregistered {} from {} index".format(
+            result_string, index_sha
+        ))
 
 
 def get_profile_list_for_minor(base_dir, minor_version):