+++ perun/postprocess/clusterizer/run.py
@@ -20,8 +20,8 @@ import perun.utils as utils
 import perun.utils.log as log
 import perun.logic.runner as runner
 
-from perun.utils.helpers import PostprocessStatus
 from perun.profile.factory import pass_profile
+from perun.utils.structs import PostprocessStatus
 
 
 def resource_sort_key(resource):