+++ perun/utils/decorators.py
@@ -5,9 +5,6 @@ the functions. Or various checker function, that checks given parameters of the 
 """
 
 import inspect
-import time
-
-import termcolor
 
 from perun.utils.exceptions import InvalidParameterException
 
@@ -158,32 +155,4 @@ def phase_function(phase_name):
         """
         func.phase_name = phase_name
         return func
-    return inner_wrapper
-
-
-def print_elapsed_time(func):
-    """Prints elapsed time after the execution of the wrapped function
-
-    Takes the timestamp before the execution of the function and after the execution and prints
-    the elapsed time to the standard output.
-
-    :param function func: wrapped function
-    :return: function for which we will print the elapsed time
-    """
-    def inner_wrapper(*args, **kwargs):
-        """Inner wrapper of the decorated function
-
-        :param list args: original arguments of the function
-        :param dict kwargs: original keyword arguments of the function
-        :return: results of the decorated function
-        """
-        before = time.time()
-        results = func(*args, **kwargs)
-        elapsed = time.time() - before
-        print("[!] {} [{}] in {} [!]".format(
-            (func.phase_name if hasattr(func, 'phase_name') else func.__name__).title(),
-            termcolor.colored("DONE", 'green', attrs=['bold']),
-            termcolor.colored("{:0.2f}s".format(elapsed), 'white', attrs=['bold'])
-        ))
-        return results
     return inner_wrapper