+++ perun/utils/log.py
@@ -11,17 +11,19 @@ import pydoc
 import functools
 import traceback
 import numpy as np
+import time
 
 import termcolor
 
-from perun.utils.helpers import first_index_of_attr
+from perun.utils.helpers import first_index_of_attr, str_to_plural
 from perun.utils.decorators import static_variables
-from perun.utils.helpers import COLLECT_PHASE_ATTRS, COLLECT_PHASE_ATTRS_HIGH, CHANGE_COLOURS, \
-    CHANGE_STRINGS, DEGRADATION_ICON, OPTIMIZATION_ICON, CHANGE_CMD_COLOUR, CHANGE_TYPE_COLOURS
+from perun.utils.helpers import COLLECT_PHASE_ATTRS, CHANGE_COLOURS, CHANGE_STRINGS, \
+    DEGRADATION_ICON, OPTIMIZATION_ICON, CHANGE_CMD_COLOUR, CHANGE_TYPE_COLOURS
 from perun.utils.structs import PerformanceChange
 
 __author__ = 'Tomas Fiedor'
 VERBOSITY = 0
+COLOR_OUTPUT = True
 
 # Enum of verbosity levels
 VERBOSE_DEBUG = 2
@@ -130,7 +132,7 @@ def info(msg):
     """
     :param str msg: info message that will be printed only when there is at least lvl1 verbosity
     """
-    print("info: {}".format(msg))
+    print("{}".format(msg))
 
 
 def quiet_info(msg):
@@ -173,7 +175,7 @@ def print_current_stack(colour='red'):
         ]
         if not any(filtering_conditions):
             reduced_trace.append(frame)
-    print(termcolor.colored(
+    print(in_color(
         ''.join(traceback.format_list(reduced_trace)), colour
     ), file=sys.stderr)
 
@@ -183,7 +185,7 @@ def error(msg, recoverable=False):
     :param str msg: error message printed to standard output
     :param bool recoverable: whether we can recover from the error
     """
-    print(termcolor.colored("fatal: {}".format(msg), 'red'), file=sys.stderr)
+    print(in_color("fatal: {}".format(msg), 'red'), file=sys.stderr)
     if is_verbose_enough(VERBOSE_DEBUG):
         print_current_stack()
 
@@ -198,7 +200,7 @@ def warn(msg, end="\n"):
     :param str end:
     """
     if not SUPPRESS_WARNINGS:
-        print("warn: {}".format(msg), end=end)
+        print("warning: {}".format(msg), end=end)
 
 
 def print_current_phase(phase_msg, phase_unit, phase_colour):
@@ -208,10 +210,8 @@ def print_current_phase(phase_msg, phase_unit, phase_colour):
     :param str phase_unit: additional parameter that is passed to the phase_msg
     :param str phase_colour: phase colour defined in helpers.py
     """
-    print(termcolor.colored(
-        phase_msg.format(
-            termcolor.colored(phase_unit, attrs=COLLECT_PHASE_ATTRS_HIGH)
-        ), phase_colour, attrs=COLLECT_PHASE_ATTRS
+    print(in_color(
+        phase_msg.format(in_color(phase_unit)), phase_colour, COLLECT_PHASE_ATTRS
     ))
 
 
@@ -228,28 +228,26 @@ def print_job_progress(overall_jobs):
     print_job_progress.current_job += 1
 
 
-def cprint(string, colour, attrs=None, flush=True):
+def cprint(string, colour, attrs='none', flush=True):
     """Wrapper over coloured print without adding new line
 
     :param str string: string that is printed with colours
     :param str colour: colour that will be used to colour the string
-    :param list attrs: list of additional attributes for the colouring
+    :param str attrs: name of additional attributes for the colouring
     :param bool flush: set True to immediately perform print operation
     """
-    attrs = attrs or []
-    print(termcolor.colored(string, colour, attrs=attrs), end='', flush=flush)
+    print(in_color(string, colour, attrs), end='', flush=flush)
 
 
-def cprintln(string, colour, attrs=None, ending='\n'):
+def cprintln(string, colour, attrs='none'):
     """Wrapper over coloured print with added new line or other ending
 
     :param str string: string that is printed with colours and newline
     :param str colour: colour that will be used to colour the stirng
-    :param list attrs: list of additional attributes for the colouring
+    :param str attrs: name of additional attributes for the colouring
     :param str ending: ending of the string, be default new line
     """
-    attrs = attrs or []
-    print(termcolor.colored(string, colour, attrs=attrs), end=ending)
+    print(in_color(string, colour, attrs))
 
 
 def done(ending='\n'):
@@ -258,7 +256,7 @@ def done(ending='\n'):
     :param str ending: end of the string, by default new line
     """
     print('[', end='')
-    cprint("DONE", 'green', attrs=['bold'])
+    cprint("DONE", 'green', attrs='bold')
     print(']', end=ending)
 
 
@@ -267,23 +265,29 @@ def failed(ending='\n'):
     :param str ending: end of the string, by default new line
     """
     print('[', end='')
-    cprint("FAILED", 'red', attrs=['bold'])
+    cprint("FAILED", 'red', attrs='bold')
     print(']', end=ending)
 
 
-def set_color(output, color, enable_coloring=True, attrs=None):
+def in_color(output, color='white', attribute_style="none"):
     """Transforms the output to colored version.
 
     :param str output: the output text that should be colored
     :param str color: the color
-    :param bool enable_coloring: switch that allows to disable the coloring - the function is no-op
-    :param list attrs: list of additional attributes for the coloring
+    :param str attribute_style: name of the additional style, i.e. bold, italic, etc.
 
     :return str: the new colored output (if enabled)
     """
-    if enable_coloring:
+    attrs = {
+        "none": [],
+        "bold": ["bold"],
+        "underline": ["underline"]
+    }.get(attribute_style, [])
+
+    if COLOR_OUTPUT:
         return termcolor.colored(output, color, attrs=attrs)
-    return output
+    else:
+        return output
 
 
 def count_degradations_per_group(degradation_list):
@@ -337,9 +341,9 @@ def print_short_summary_of_degradations(degradation_list):
     print_short_change_string(counts)
     optimization_count = counts.get('Optimization', 0)
     degradation_count = counts.get('Degradation', 0)
-    print("{} optimization{}({}), {} degradation{}({})".format(
-        optimization_count, "s" if optimization_count != 1 else "", OPTIMIZATION_ICON,
-        degradation_count, "s" if degradation_count != 1 else "", DEGRADATION_ICON
+    print("{}({}), {}({})".format(
+        str_to_plural(optimization_count, "optimization"), OPTIMIZATION_ICON,
+        str_to_plural(degradation_count, "degradation"), DEGRADATION_ICON
     ))
 
 
@@ -351,15 +355,15 @@ def change_counts_to_string(counts, width=0):
     :return: string representing the counts of found changes
     """
     width = max(width - counts.get('Optimization', 0) - counts.get('Degradation', 0), 0)
-    change_str = termcolor.colored(
+    change_str = in_color(
         str(OPTIMIZATION_ICON*counts.get('Optimization', 0)),
         CHANGE_COLOURS[PerformanceChange.Optimization],
-        attrs=['bold']
+        'bold'
     )
-    change_str += termcolor.colored(
+    change_str += in_color(
         str(DEGRADATION_ICON*counts.get('Degradation', 0)),
         CHANGE_COLOURS[PerformanceChange.Degradation],
-        attrs=['bold']
+        'bold'
     )
     return change_str + width*' '
 
@@ -376,9 +380,7 @@ def print_short_change_string(counts):
     :param dict counts: dictionary mapping found string changes into their counts
     """
     overall_changes = sum(counts.values())
-    print("{} change{}".format(
-        overall_changes, "s" if overall_changes != 1 else ""
-    ), end='')
+    print(str_to_plural(overall_changes, "change"), end='')
     if overall_changes > 0:
         change_string = change_counts_to_string(counts)
         print(" | {}".format(change_string), end='')
@@ -411,7 +413,7 @@ def _print_models_info(deg_info, model_strategy):
         :param str baseline_colour: baseline colour to print baseline string
         :param str target_str: target kind of model (e.g. moving_average, linear, etc.)
         :param str target_colour: target colour to print target string
-        :param list attrs: list of additional attributes for the colouring
+        :param str attrs: name of type attributes for the colouring
         :return None: function has nor return value
         """
         print(baseline_str, end='')
@@ -422,19 +424,19 @@ def _print_models_info(deg_info, model_strategy):
     from_colour, to_colour = get_degradation_change_colours(deg_info.result)
 
     if model_strategy == 'best-param':
-        print_models_kinds(' from: ', from_colour, ' -> to: ', to_colour, ['bold'])
+        print_models_kinds(' from: ', from_colour, ' -> to: ', to_colour, 'bold')
     elif model_strategy in ('best-nonparam', 'best-model', 'best-both'):
-        print_models_kinds(' base: ', 'blue', ' targ: ', 'blue', ['bold'])
+        print_models_kinds(' base: ', 'blue', ' targ: ', 'blue', 'bold')
     elif model_strategy in ('all-nonparam', 'all-param', 'all-models'):
         print(' model: ', end='')
-        cprint('{}'.format(deg_info.from_baseline), colour='blue', attrs=['bold'])
+        cprint('{}'.format(deg_info.from_baseline), colour='blue', attrs='bold')
 
     if deg_info.confidence_type != 'no':
         print(' (with confidence ', end='')
         cprint(
             '{} = {}'.format(
                 deg_info.confidence_type, deg_info.confidence_rate),
-            'white', attrs=['bold']
+            'white', 'bold'
         )
         print(')', end='')
 
@@ -456,7 +458,7 @@ def _print_partial_intervals(partial_intervals):
         if change_info != PerformanceChange.NoChange:
             cprint(
                 "<{}, {}> {}x; ".format(x_start, x_end, rel_error),
-                CHANGE_COLOURS.get(change_info, 'white'), attrs=[]
+                CHANGE_COLOURS.get(change_info, 'white')
             )
     print("")
 
@@ -485,25 +487,25 @@ def print_list_of_degradations(degradation_list, model_strategy="best-model"):
     for location, changes in itertools.groupby(degradation_list, keygetter):
         # Print the location
         print("at", end='')
-        cprint(' {}'.format(location), 'white', attrs=['bold'])
+        cprint(' {}'.format(location), 'white', attrs='bold')
         print(":")
         # Iterate and print all of the infos
         for deg_info, cmd, __ in changes:
             print('\u2514 ', end='')
-            cprint('{}x'.format(round(deg_info.rate_degradation, 2)), 'white', attrs=['bold'])
+            cprint('{}x'.format(round(deg_info.rate_degradation, 2)), 'white', 'bold')
             print(': ', end='')
-            cprint(deg_info.type, CHANGE_TYPE_COLOURS.get(deg_info.type, 'white'), attrs=[])
+            cprint(deg_info.type, CHANGE_TYPE_COLOURS.get(deg_info.type, 'white'))
             print(' ', end='')
             cprint(
                 '{}'.format(CHANGE_STRINGS[deg_info.result]),
-                CHANGE_COLOURS[deg_info.result], attrs=['bold']
+                CHANGE_COLOURS[deg_info.result], 'bold'
             )
             if deg_info.result != PerformanceChange.NoChange:
                 _print_models_info(deg_info, model_strategy)
 
             # Print information about command that was executed
             print(" (", end='')
-            cprint("$ {}".format(cmd), CHANGE_CMD_COLOUR, attrs=['bold'])
+            cprint("$ {}".format(cmd), CHANGE_CMD_COLOUR, 'bold')
             print(')')
 
             # Print information about the change on the partial intervals (only at Local-Statistics)
@@ -566,6 +568,33 @@ def aggregate_intervals(intervals):
 
     return agg_intervals
 
+def print_elapsed_time(func):
+    """Prints elapsed time after the execution of the wrapped function
+
+    Takes the timestamp before the execution of the function and after the execution and prints
+    the elapsed time to the standard output.
+
+    :param function func: wrapped function
+    :return: function for which we will print the elapsed time
+    """
+    def inner_wrapper(*args, **kwargs):
+        """Inner wrapper of the decorated function
+
+        :param list args: original arguments of the function
+        :param dict kwargs: original keyword arguments of the function
+        :return: results of the decorated function
+        """
+        before = time.time()
+        results = func(*args, **kwargs)
+        elapsed = time.time() - before
+        print("[!] {} [{}] in {} [!]".format(
+            (func.phase_name if hasattr(func, 'phase_name') else func.__name__).title(),
+            in_color("DONE", 'green', 'bold'),
+            in_color("{:0.2f}s".format(elapsed), 'white', 'bold')
+        ))
+        return results
+    return inner_wrapper
+
 
 class History:
     """Helper with wrapper, which is used when one wants to visualize the version control history
@@ -604,7 +633,7 @@ class History:
             :return: string representing the edge in ascii
             """
             return char if self.colour == 'white' \
-                else termcolor.colored(char, self.colour, attrs=['bold'])
+                else in_color(char, self.colour, 'bold')
 
     def __init__(self, head):
         """Creates a with wrapper, which keeps and prints the context of the current vcs
@@ -717,7 +746,7 @@ class History:
         print(minor_str, end='')
         cprint(" {}".format(
             minor_version_info.checksum[:6]
-        ), 'yellow', attrs=[])
+        ), 'yellow')
         print(": {} | ".format(
             minor_version_info.desc.split("\n")[0].strip()
         ), end='')