+++ tests/test_collect.py
@@ -3,6 +3,8 @@
 import os
 import glob
 import re
+import subprocess
+import signal
 
 from click.testing import CliRunner
 
@@ -19,8 +21,9 @@ import perun.collect.complexity.symbols as symbols
 import perun.collect.complexity.run as complexity
 import perun.utils.log as log
 
+from perun.profile.factory import Profile
 from perun.utils.helpers import Job
-from perun.utils.structs import Unit
+from perun.utils.structs import Unit, Executable, CollectStatus, RunnerReport
 from perun.workload.integer_generator import IntegerGenerator
 from perun.collect.trace.systemtap import _TraceRecord
 from perun.collect.trace.systemtap_script import RecordType
@@ -196,7 +199,7 @@ def test_collect_complexity(monkeypatch, helpers, pcs_full, complexity_collect_j
     cmd, args, work, collectors, posts, config = complexity_collect_job
     head = vcs.get_minor_version_info(vcs.get_minor_head())
     result = run.run_single_job(cmd, args, work, collectors, posts, [head], **config)
-    assert result == run.CollectStatus.OK
+    assert result == CollectStatus.OK
 
     # Assert that nothing was removed
     after_object_count = helpers.count_contents_on_path(pcs_full.get_path())[0]
@@ -249,9 +252,15 @@ def test_collect_complexity(monkeypatch, helpers, pcs_full, complexity_collect_j
     assert result.exit_code == 0
     assert 'stored profile' in result.output
 
-    monkeypatch.setattr(
-        "perun.utils.build_command_str", lambda *_: "nonexistent"
-    )
+    original_run = utils.run_safely_external_command
+    def patched_run(cmd, *args, **kwargs):
+        if cmd.startswith('g++') or cmd.startswith('readelf') or cmd.startswith('echo'):
+            return original_run(cmd, *args, **kwargs)
+        else:
+            raise subprocess.CalledProcessError(1, 'error')
+
+    monkeypatch.setattr('perun.utils.run_safely_external_command', patched_run)
+
     runner = CliRunner()
     result = runner.invoke(cli.collect, ['-c{}'.format(job_params['target_dir']),
                                          '-a test', '-w input', 'complexity',
@@ -598,8 +607,11 @@ def test_collect_memory(capsys, helpers, pcs_full, memory_collect_job, memory_co
         'all': False,
         'no_func': 'main'
     })
-    job = Job('memory', [], str(target_bin), '', '')
+    executable = Executable(str(target_bin))
+    assert executable.to_escaped_string() != ""
+    job = Job('memory', [], executable)
     _, prof = run.run_collector(collector_unit, job)
+    prof = Profile(prof)
 
     assert len(list(prof.all_resources())) == 2
 
@@ -607,17 +619,18 @@ def test_collect_memory(capsys, helpers, pcs_full, memory_collect_job, memory_co
         'all': False,
         'no_source': 'memory_collect_test.c'
     })
-    job = Job('memory', [], str(target_bin), '', '')
+    job = Job('memory', [], executable)
     _, prof = run.run_collector(collector_unit, job)
+    prof = Profile(prof)
 
     assert len(list(prof.all_resources())) == 0
 
 
 def test_collect_memory_with_generator(pcs_full, memory_collect_job):
     """Tries to collect the memory with integer generators"""
-    cmd = memory_collect_job[0][0]
+    executable = Executable(memory_collect_job[0][0])
     collector = Unit('memory', {})
-    integer_job = Job(collector, [], cmd, '', '')
+    integer_job = Job(collector, [], executable)
     integer_generator = IntegerGenerator(integer_job, 1, 3, 1)
     memory_profiles = list(integer_generator.generate(run.run_collector))
     assert len(memory_profiles) == 1
@@ -656,3 +669,69 @@ def test_collect_time(monkeypatch, helpers, pcs_full, capsys):
     run.run_single_job(["echo"], "", ["hello"], ["time"], [], [head])
     _, err = capsys.readouterr()
     assert 'Something happened lol!' in err
+
+
+def test_integrity_tests(capsys):
+    """Basic tests for checking integrity of runners"""
+    mock_report = RunnerReport(complexity, 'postprocessor', {'profile': {}})
+    run.check_integrity_of_runner(complexity, 'postprocessor', mock_report)
+    out, err = capsys.readouterr()
+    assert "warning: complexity is missing postprocess() function" in out
+    assert "" == err
+
+    mock_report = RunnerReport(complexity, 'collector', {})
+    run.check_integrity_of_runner(complexity, 'collector', mock_report)
+    out, err = capsys.readouterr()
+    assert "warning: collector complexity does not return any profile"
+    assert "" == err
+
+
+def test_teardown(pcs_full, monkeypatch, capsys):
+    """Basic tests for integrity of the teardown phase"""
+    head = vcs.get_minor_version_info(vcs.get_minor_head())
+    original_phase_f = run.run_phase_function
+
+    # Assert that collection went OK and teardown returns error
+    status = run.run_single_job(["echo"], "", ["hello"], ["time"], [], [head])
+    assert status == CollectStatus.OK
+
+    def teardown_returning_error(report, phase):
+        if phase == 'teardown':
+            result = (CollectStatus.ERROR, "error in teardown", {})
+            report.update_from(*result)
+        else:
+            original_phase_f(report, phase)
+    monkeypatch.setattr("perun.logic.runner.run_phase_function", teardown_returning_error)
+    status = run.run_single_job(["echo"], "", ["hello"], ["time"], [], [head])
+    assert status == CollectStatus.ERROR
+    _, err = capsys.readouterr()
+    assert "error in teardown" in err
+
+    # Assert that collection went Wrong, teardown went OK, and still returns the error
+    def teardown_not_screwing_things(report, phase):
+        if phase == 'teardown':
+            print("Teardown was executed")
+            assert report.status == CollectStatus.ERROR
+            result = (CollectStatus.OK, "", {})
+            report.update_from(*result)
+        elif phase == 'before':
+            result = (CollectStatus.ERROR, "error while before", {})
+            report.update_from(*result)
+    monkeypatch.setattr("perun.logic.runner.run_phase_function", teardown_not_screwing_things)
+    status = run.run_single_job(["echo"], "", ["hello"], ["time"], [], [head])
+    assert status == CollectStatus.ERROR
+    out, err = capsys.readouterr()
+    assert "Teardown was executed" in out
+    assert "error while before" in err
+
+    # Test signals
+    def collect_firing_sigint(report, phase):
+        if phase == 'collect':
+            os.kill(os.getpid(), signal.SIGINT)
+        else:
+            original_phase_f(report, phase)
+    monkeypatch.setattr("perun.logic.runner.run_phase_function", collect_firing_sigint)
+    status = run.run_single_job(["echo"], "", ["hello"], ["time"], [], [head])
+    assert status == CollectStatus.ERROR
+    out, err = capsys.readouterr()
+    assert "fatal: while collecting by time: Received signal: 2" in err