+++ CHANGELOG.rst
@@ -6,6 +6,18 @@ HEAD
 
 **To be included in next release**
 
+  - extract cmd, args and workload to Executable class
+  - remove ``--remove-all`` argument in ``perun rm``
+  - add support for removing profiles from pending jobs through perun
+  - improve the output of `perun rm` command
+  - extract CLI groups to isolate modules
+  - add caching to selected vcs commands
+  - fix untested bug in degradation check
+  - rename warmup parameter in `time` to ``--warmup``
+  - lower the number of warmup and repetitions for time collector during tests
+  - remove filter postprocessor (did nothing)
+  - add signal handling to runner (authored by Jirka Pavela)
+
 0.17.1 (2019-07-24)
 -------------------